refactor(dialog): extract useNativeDialog hook
Centralize native <dialog> showModal/close and cancel handling so
ConfirmDialog and other modals share one lifecycle implementation.
Add configurable id and submittingText props for dialog variants.
